chore: Update Angular project configuration and dependencies to no standalone
