+++ src/app/coming-soon-page/coming-soon-page.component.ts
@@ -2,8 +2,6 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-coming-soon-page',
-  standalone: true,
-  imports: [],
   templateUrl: './coming-soon-page.component.html',
   styleUrl: './coming-soon-page.component.css'
 })
