--- src/app/coming-soon-page/comingSoon.routes.ts
@@ -1,9 +1,0 @@
-import { Route } from "@angular/router";
-import { ComingSoonPageComponent } from "./coming-soon-page.component";
-
-export const ComingSoonRoutes: Route[] = [
-    {
-        path: '',
-        component: ComingSoonPageComponent
-    }
-]
