+++ src/app/landing-page/landing-page.component.ts
@@ -2,8 +2,6 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-landing-page',
-  standalone: true,
-  imports: [],
   templateUrl: './landing-page.component.html',
   styleUrl: './landing-page.component.css'
 })
