+++ src/environments/environment.ts
@@ -1,4 +1,3 @@
-
 export const environment = {
     firebase: {
         apiKey: "AIzaSyBiy2Abb6OPUDEmGvW01ZQJaaKhQbfyU-c",
